继续增删改查

+++ springH/src/main/com/lx/services/UserService.java
@@ -4,5 +4,5 @@ import com.lx.models.User;
 import com.lx.utils.PageData;
 
 public interface UserService extends Service<User, Long>{
-	public PageData<User> serchUsers(User user,Long currentPage);
+	public PageData<User> serchUsers(User user,int currentPage);
 }
